fix/pay-os

+++ src/PerfectBreakfast.Application/Repositories/IOrderRepository.cs
@@ -5,7 +5,7 @@ namespace PerfectBreakfast.Application.Repositories
 {
     public interface IOrderRepository : IGenericRepository<Order>
     {
-        public Task<Order> GetOrderByOrderCode(int orderCode);
+        public Task<Order> GetOrderByOrderCode(long orderCode);
         public Task<List<Order>> GetOrderByDailyOrderId(Guid dailyOrderId);
         public Task<List<Order>> GetOrderHistory(Guid userId, params IncludeInfo<Order>[] includeProperties);
     }
